Add external service to subscribe to achievement events and update in real time
