+++ services/achievements-listener/tsconfig.json
@@ -1,0 +1,24 @@
+{
+  "compilerOptions": {
+    "target": "ES2022",
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
+    "moduleDetection": "force",
+    "allowImportingTsExtensions": true,
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "types": ["node"],
+    "noEmit": true,
+    "baseUrl": ".",
+    "paths": {
+      "@tailtag/achievements-processor": ["../../supabase/shared/achievements/processor.ts"]
+    }
+  },
+  "include": [
+    "src/**/*.ts",
+    "../../supabase/shared/**/*.ts"
+  ]
+}
